Refactored JPanels to panel package
