+++ NineMenMorris/src/main/java/projectmanagers/gui/panels/Player1Panel.java
@@ -1,15 +1,16 @@
-package main.java.projectmanagers.gui;
+package main.java.projectmanagers.gui.panels;
+
 import javax.swing.*;
 import java.awt.*;
 
-public class Player2Panel extends JPanel {
-    public JLabel player2Txt;
+public class Player1Panel extends JPanel {
+    public JLabel player1Txt;
     private int turns = 9;
     private GridBagConstraints gbc;
 
-    public Player2Panel () {
+    public Player1Panel () {
         gbc = new GridBagConstraints();
-        player2Txt = new JLabel("Player 2");
+        player1Txt = new JLabel("Player 1");
         buildPanel();
     }
     public void buildPanel () {
@@ -17,8 +18,8 @@ public class Player2Panel extends JPanel {
         setLayout(new GridBagLayout());
         setBackground(new Color(116,101,72));
         setPreferredSize(new Dimension(75,600));
-        player2Txt.setHorizontalAlignment(SwingConstants.CENTER);
-        add(player2Txt, gbc);
+        player1Txt.setHorizontalAlignment(SwingConstants.CENTER);
+        add(player1Txt, gbc);
         trackTurns();
     }
     public void trackTurns () {
@@ -27,3 +28,4 @@ public class Player2Panel extends JPanel {
         }
     }
 }
+
