+++ NineMenMorris/src/main/java/projectmanagers/gui/panels/Player2Panel.java
@@ -1,16 +1,15 @@
-package main.java.projectmanagers.gui;
-
+package main.java.projectmanagers.gui.panels;
 import javax.swing.*;
 import java.awt.*;
 
-public class Player1Panel extends JPanel {
-    public JLabel player1Txt;
+public class Player2Panel extends JPanel {
+    public JLabel player2Txt;
     private int turns = 9;
     private GridBagConstraints gbc;
 
-    public Player1Panel () {
+    public Player2Panel () {
         gbc = new GridBagConstraints();
-        player1Txt = new JLabel("Player 1");
+        player2Txt = new JLabel("Player 2");
         buildPanel();
     }
     public void buildPanel () {
@@ -18,8 +17,8 @@ public class Player1Panel extends JPanel {
         setLayout(new GridBagLayout());
         setBackground(new Color(116,101,72));
         setPreferredSize(new Dimension(75,600));
-        player1Txt.setHorizontalAlignment(SwingConstants.CENTER);
-        add(player1Txt, gbc);
+        player2Txt.setHorizontalAlignment(SwingConstants.CENTER);
+        add(player2Txt, gbc);
         trackTurns();
     }
     public void trackTurns () {
@@ -28,4 +27,3 @@ public class Player1Panel extends JPanel {
         }
     }
 }
-
